split favorites
remove confusing names for state and models
