+++ src/favorites-manager/components/FavoritesBulkLoader.tsx
@@ -1,5 +1,5 @@
-import { LoadingAnimation } from "../components/LoadingAnimation";
-import { useFavoritesBatchInProgress } from "./state.manage";
+import { LoadingAnimation } from "../../components/LoadingAnimation";
+import { useFavoritesBatchInProgress } from "../../favorites/state";
 
 export const FavoritesBulkLoader = () => {
   const isLoading = useFavoritesBatchInProgress();
