+++ src/favorites/model.ts
