+++ src/recipes/components/RecipeCount.tsx
@@ -1,4 +1,4 @@
-import { useRecipeList } from "./state";
+import { useRecipeList } from "../state";
 
 export const RecipeCount = () => {
   const count = useRecipeList().count;
