+++ src/recipes/components/RecipeLoader.tsx
@@ -1,5 +1,5 @@
-import { LoadingAnimation } from "../components/LoadingAnimation";
-import { useIsLoadingRecipes } from "./state";
+import { LoadingAnimation } from "../../components/LoadingAnimation";
+import { useIsLoadingRecipes } from "../state";
 
 export const RecipeLoader = () => {
   const isLoading = useIsLoadingRecipes();
